+++ tfhe/tlwe.py
@@ -36,17 +36,17 @@ class TLweSampleArray:
         self.a = TorusPolynomialArray(params.N, shape + (self.k + 1,))
 
         # avg variance of the sample
-        self.current_variances = numpy.zeros(shape, numpy.float64)
+        self.current_variances = numpy.zeros(shape, Float)
 
         self.shape = shape
 
     def to_gpu(self, thr):
         self.a.to_gpu(thr)
-        self.current_variances = thr.to_device(self.current_variances)
+        self.current_variances = thr.to_device(self.current_variances.astype(Float))
 
     def from_gpu(self):
         self.a.from_gpu()
-        self.current_variances = self.current_variances.get()
+        self.current_variances = self.current_variances.get().astype(numpy.float64)
 
 
 class TLweSampleFFTArray:
@@ -58,17 +58,17 @@ class TLweSampleFFTArray:
         self.a = LagrangeHalfCPolynomialArray(params.N, shape + (self.k + 1,))
 
         # avg variance of the sample
-        self.current_variances = numpy.zeros(shape, numpy.float64)
+        self.current_variances = numpy.zeros(shape, Float)
 
         self.shape = shape
 
     def to_gpu(self, thr):
         self.a.to_gpu(thr)
-        self.current_variances = thr.to_device(self.current_variances)
+        self.current_variances = thr.to_device(self.current_variances.astype(Float))
 
     def from_gpu(self):
         self.a.from_gpu()
-        self.current_variances = self.current_variances.get()
+        self.current_variances = self.current_variances.get().astype(numpy.float64)
